implementation of FFT radix2

+++ include/matrix.h
@@ -11,7 +11,7 @@
      * @param  n: number of DFT points 
      * @retval the DFT matrix
      */
-complex** dft_matrix(uint8_t n);
+complex** dft_matrix(uint32_t n);
 
 /**
  * @brief  multiplication of DFT matrix with an array of complex data
@@ -21,9 +21,9 @@ complex** dft_matrix(uint8_t n);
  * @param  vector_out: output vector
  * @param  n: size of the input vector
  */
-void cpx_matrix_vector_multi(complex** matrix, complex* vector_in, complex* vector_out, uint8_t n);
+void cpx_matrix_vector_multi(complex** matrix, complex* vector_in, complex* vector_out, uint32_t n);
 
 // prints the DFT matrix 
-void cpx_print_matrix(complex** matrix, uint8_t n);
+void cpx_print_matrix(complex** matrix, uint32_t n);
 
 #endif
